move css template to templates dir
