+++ src/main.js
@@ -2,7 +2,7 @@ import './fontawesome-free-5.9.0-web/css/all.min.css';
 import './assets/scss/styles.scss';
 import './scripts/color/color-changer';
 import './scripts/contrast';
-import './scripts/cssTemplate';
+import './scripts/templates/cssTemplate';
 import './scripts/icons/editIcons';
 import './scripts/copy-paste/copy-paste';
 import './scripts/router/routes';
